--story=1003067 --user=林邦泉 流程编辑页面，自动处理节点补充‘权限设置’和‘按钮映射’ https://www.tapd.cn/54247054/s/1074717

+++ src/main/java/codedriver/module/process/api/processstep/ProcessStepHandleConfigSaveApi.java
@@ -6,6 +6,8 @@ import codedriver.framework.notify.dto.NotifyPolicyInvokerVo;
 import codedriver.framework.process.constvalue.ProcessStepHandler;
 import codedriver.framework.process.dao.mapper.ProcessStepHandlerMapper;
 import codedriver.framework.process.dto.ProcessStepHandlerVo;
+import codedriver.framework.process.stephandler.core.IProcessStepHandler;
+import codedriver.framework.process.stephandler.core.ProcessStepHandlerFactory;
 import codedriver.framework.restful.annotation.Description;
 import codedriver.framework.restful.annotation.Input;
 import codedriver.framework.restful.annotation.Param;
@@ -51,33 +53,34 @@ public class ProcessStepHandleConfigSaveApi extends ApiComponentBase {
     }
 
     @Input({
-    	//@Param( name = "handler", type = ApiParamType.STRING, desc = "流程节点组件", isRequired = true),
-        //@Param( name = "config", type = ApiParamType.JSONOBJECT, desc = "流程节点组件配置", isRequired = true)
     	@Param(name = "processStepHandlerList", type = ApiParamType.JSONARRAY, isRequired = true, desc = "流程节点组件配置信息列表")
     })
     @Description(desc = "流程节点组件配置保存接口")
     @Override
     public Object myDoService(JSONObject jsonObj) throws Exception {
-//    	ProcessStepHandlerVo stepHandlerVo = JSON.parseObject(jsonObj.toJSONString(), new TypeReference<ProcessStepHandlerVo>() {});
-    	List<ProcessStepHandlerVo> processStepHandlerList = JSON.parseArray(jsonObj.getJSONArray("processStepHandlerList").toJSONString(), ProcessStepHandlerVo.class);
+    	List<ProcessStepHandlerVo> processStepHandlerList = JSON.parseArray(JSON.toJSONString(jsonObj.getJSONArray("processStepHandlerList")), ProcessStepHandlerVo.class);
     	for(ProcessStepHandlerVo stepHandlerVo :processStepHandlerList) {
-    		stepHandlerMapper.deleteProcessStepHandlerConfigByHandler(stepHandlerVo.getHandler());
-            stepHandlerMapper.insertProcessStepHandlerConfig(stepHandlerVo);
-            notifyPolicyInvokerManager.removeInvoker(stepHandlerVo.getHandler());
-            JSONObject config = stepHandlerVo.getConfig();
-            JSONObject notifyPolicyConfig = config.getJSONObject("notifyPolicyConfig");
-            Long policyId = notifyPolicyConfig.getLong("policyId");
-            if(policyId != null) {
-            	NotifyPolicyInvokerVo notifyPolicyInvokerVo = new NotifyPolicyInvokerVo();
-            	notifyPolicyInvokerVo.setPolicyId(policyId);
-            	notifyPolicyInvokerVo.setInvoker(stepHandlerVo.getHandler());
-            	JSONObject notifyPolicyInvokerConfig = new JSONObject();
-            	notifyPolicyInvokerConfig.put("function", "processstephandler");
-            	notifyPolicyInvokerConfig.put("name", "节点管理-" + ProcessStepHandler.getName(stepHandlerVo.getHandler()));
-            	notifyPolicyInvokerConfig.put("handler", stepHandlerVo.getHandler());
-            	notifyPolicyInvokerVo.setConfig(notifyPolicyInvokerConfig.toJSONString());
-            	notifyPolicyInvokerManager.addInvoker(notifyPolicyInvokerVo);
-            }
+    		IProcessStepHandler handler = ProcessStepHandlerFactory.getHandler(stepHandlerVo.getHandler());
+    		if(handler != null) {
+    			JSONObject config = handler.makeupConfig(stepHandlerVo.getConfig());
+    			stepHandlerVo.setConfig(config.toJSONString());
+    			stepHandlerMapper.deleteProcessStepHandlerConfigByHandler(stepHandlerVo.getHandler());
+                stepHandlerMapper.insertProcessStepHandlerConfig(stepHandlerVo);
+                notifyPolicyInvokerManager.removeInvoker(stepHandlerVo.getHandler());
+                JSONObject notifyPolicyConfig = config.getJSONObject("notifyPolicyConfig");
+                Long policyId = notifyPolicyConfig.getLong("policyId");
+                if(policyId != null) {
+                	NotifyPolicyInvokerVo notifyPolicyInvokerVo = new NotifyPolicyInvokerVo();
+                	notifyPolicyInvokerVo.setPolicyId(policyId);
+                	notifyPolicyInvokerVo.setInvoker(stepHandlerVo.getHandler());
+                	JSONObject notifyPolicyInvokerConfig = new JSONObject();
+                	notifyPolicyInvokerConfig.put("function", "processstephandler");
+                	notifyPolicyInvokerConfig.put("name", "节点管理-" + ProcessStepHandler.getName(stepHandlerVo.getHandler()));
+                	notifyPolicyInvokerConfig.put("handler", stepHandlerVo.getHandler());
+                	notifyPolicyInvokerVo.setConfig(notifyPolicyInvokerConfig.toJSONString());
+                	notifyPolicyInvokerManager.addInvoker(notifyPolicyInvokerVo);
+                }
+    		}   		
     	}        
         return null;
     }
